code/DHService/pyproject.toml
[project]
name = "dhservice"
version = "0.1.0"
description = "The Deep Harbor services API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.121.3",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
